casa_coordinates::coordinate

Enum CoordinateModel

pub enum CoordinateModel {
    Linear(LinearCoordinate),
    Direction(DirectionCoordinate),
    Spectral(Box<SpectralCoordinate>),
    Stokes(StokesCoordinate),
    Tabular(TabularCoordinate),
}

Closed model for every coordinate kind supported by casacore images.

Variants

Linear(LinearCoordinate)

Direction(DirectionCoordinate)

Spectral(Box<SpectralCoordinate>)

Stokes(StokesCoordinate)

Tabular(TabularCoordinate)

Implementations

impl CoordinateModel

pub fn coordinate_type(&self) -> CoordinateType

pub fn n_pixel_axes(&self) -> usize

pub fn n_world_axes(&self) -> usize

pub fn to_world(&self, pixel: &[f64]) -> Result<Vec<f64>, CoordinateError>

pub fn to_pixel(&self, world: &[f64]) -> Result<Vec<f64>, CoordinateError>

pub fn reference_value(&self) -> Vec<f64>

pub fn reference_pixel(&self) -> Vec<f64>

pub fn increment(&self) -> Vec<f64>

pub fn axis_names(&self) -> Vec<String>

pub fn axis_units(&self) -> Vec<String>

pub fn to_record(&self) -> RecordValue

Serializes this coordinate in the canonical casacore field layout.